Resolve issues about authors and propertie spdx in license. Fixes #741, #744, #745
