+++ src/somef/utils/constants.py
@@ -44,12 +44,12 @@ REGEXP_TITLE_NATURAL = r'["“](.+?)["”]'
 REGEXP_APACHE = r'(?i)apache\s+license\s*,?\s*version\s*2\.0'
 REGEXP_GPL3 = r'(?i)gnu\s+general\s+public\s+license\s*,?\s*version\s*3\.0'
 REGEXP_MIT = r'(?i)mit\s+license'
-REGEXP_BSD2 = r'(?i)bsd\s+2-clause\s+license'
+REGEXP_BSD2 = r'(?i)(bsd\s*-?\s*2-?clause(?:\s*license)?|redistribution\s+and\s+use\s+in\s+source\s+and\s+binary\s+forms)'
 REGEXP_BSD3 = r'(?i)bsd\s+3-clause\s+license'
 REGEXP_BOOST = r'(?i)boost\s+software\s+license\s*,?\s*version\s*1\.0'
 REGEXP_CC0 = r'(?i)creative\s+commons\s+zero\s+v?1\.0\s+universal'
 REGEXP_EPL2 = r'(?i)eclipse\s+public\s+license\s*,?\s*version\s*2\.0'
-REGEXP_AGPL3 = r'(?i)gnu\s+affero\s+general\s+public\s+license\s*,?\s*version\s*3\.0'
+REGEXP_AGPL3 = r'(?i)gnu\s+affero\s+general\s+public\s+license\s*(?:v(?:ersion)?\.?\s*3(?:\.0)?)'
 REGEXP_GPL2 = r'(?i)gnu\s+general\s+public\s+license\s*,?\s*version\s*2\.0'
 REGEXP_LGPL1 = r'(?i)gnu\s+lesser\s+general\s+public\s+license\s*,?\s*version\s*1\.0'
 REGEXP_MPL2 = r'(?i)mozilla\s+public\s+license\s*,?\s*version\s*2\.0'
